\name{OzoneExposure-package}
\alias{OzoneExposure-package}
\alias{OzoneExposure}
\docType{package}
\title{
\packageTitle{OzoneExposure}
}
\description{
\packageDescription{OzoneExposure}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{OzoneExposure}
\packageIndices{OzoneExposure}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{OzoneExposure}

Maintainer: \packageMaintainer{OzoneExposure}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
